Normalize pace fields and AI coaching output

+++ src/main/java/com/neostride/server/coaching/repository/GoalInsertCommand.java
@@ -7,6 +7,6 @@ public record GoalInsertCommand(
 		Integer durationWeeks,
 		Integer runningDay,
 		BigDecimal targetDistance,
-		BigDecimal targetPace
+		Integer targetPace
 ) {
 }
